refactor: decouple monolithic time calculator
Extracts mathematical utilities, profiling logic, and metadata analysis from the large timeCalculator.ts file into modular files under src/utils/timeCalculator/ to improve readability and maintainability.
